Custom Scripting - Adds buttons to allow the deletion of custom scripts

+++ src/main/java/sirius/biz/scripting/mongo/MongoCustomScriptController.java
@@ -76,7 +76,9 @@ public class MongoCustomScriptController extends BizController {
     @Routed("/scripting/scripts/:1/delete")
     @Permission(ScriptingController.PERMISSION_SCRIPTING)
     public void deleteScript(WebContext webContext, String id) {
-        deleteEntity(webContext, tryFindForTenant(MongoCustomScript.class, id));
+        if (webContext.isSafePOST()) {
+            deleteEntity(webContext, tryFindForTenant(MongoCustomScript.class, id));
+        }
         webContext.respondWith().redirectToGet("/scripting/scripts");
     }
